Foundational/API/app/requirements.txt
fastapi
uvicorn
pymongo
pymupdf
python-docx
pandas
faiss-cpu
python-multipart
openai>=1.0.0
python-dotenv
feedparser